src/test/java/tomocomd/startpep/aggregtions/NormsAggParamTest.java
package tomocomd.startpep.aggregtions;

import static org.junit.jupiter.api.Assertions.*;

import java.util.Objects;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import tomocomd.peptides.dcs.startpep.aggregtions.NormsAggParam;

class NormsAggParamTest {

  NormsAggParam normsAggParam;

  @BeforeEach
  void setUp() {
    normsAggParam = new NormsAggParam();
  }

  @Test
  void setDefault() {
    assertArrayEquals(getDefaults(), normsAggParam.getValues());
  }

  @Test
  void getParamName() {
    assertEquals("Norms aggregators", normsAggParam.getParamName());
  }

  @Test
  void testValidate() {
    NormsAggParam norms = new NormsAggParam(normsAggParam);
    norms.getParams().keySet().forEach(key -> norms.setValue(key, false));
    String[] values = norms.getValues();
    assertTrue(Objects.nonNull(values) && values.length == 0);
  }

  String[] getDefaults() {
    return new String[] {"N1", "N2", "N3"};
  }
}
